migrate HostSeeker to seeker pkg
